remove unused llmconfig
