+++ src/lightspeed_evaluation/core/config/__init__.py
@@ -1,7 +1,7 @@
 """Configuration loading and models for Evaluation Framework."""
 
 from .loader import ConfigLoader, setup_environment_variables, SystemConfig
-from .models import EvaluationData, EvaluationResult, TurnData
+from .models import EvaluationData, EvaluationResult, TurnData, LLMConfig
 from .validator import DataValidator
 
 __all__ = [
@@ -11,5 +11,6 @@ __all__ = [
     "EvaluationData",
     "EvaluationResult",
     "TurnData",
+    "LLMConfig",
     "DataValidator",
 ]
