+++ src/lightspeed_evaluation/core/llm/__init__.py
@@ -1,8 +1,9 @@
 """LLM management for Evaluation Framework."""
 
-from .manager import LLMManager, LLMConfig, LLMError
+from .manager import LLMManager, LLMError
 from .deepeval import DeepEvalLLMManager
 from .ragas import RagasLLMManager
+from ..config import LLMConfig
 
 __all__ = [
     "LLMManager",
