+++ src/lightspeed_evaluation/core/metrics/deepeval.py
@@ -10,7 +10,7 @@ from deepeval.metrics import (
 from deepeval.test_case import ConversationalTestCase
 from deepeval.test_case import Turn as DeepEvalTurn
 
-from ..config.models import TurnData
+from ..config import TurnData
 from ..llm.deepeval import DeepEvalLLMManager
 from ..llm.manager import LLMManager
 from ..output.statistics import EvaluationScope
